pyproject.toml
[project]
name = "dialogues"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ollama>=0.4.7",
    "python-decouple>=3.8",
]
